Eliminate redundant _config suffix in config files
Previously config types would only be bound without a qualifier if they
had a _config suffix (e.g. web_config), which is redundant and
unnecessary given that we're in a config file. Eliminate this
requirement and allow default binding when the proper matches the "raw"
name of the configuration type (e.g. web).