Expand upon routing rules for MediaRanges
Prior to this change if we had two routes with the same path
the one returned was arbitrary, even if one was the clear choice
in regards to content type. We now attempt to choose the route which
matches the desired content type.

We also attempt to route based, in some cases, on whether or not
the specified charset matches our route.

Along with this comes two small fixes:
  - Allow multiple media ranges to be specified in the accept header
  - Parse the URL only once per request