Fix ActionScoped to avoid recursive computes
Computing the value of an ActionScoped might require computing the
result of other ActionScoped objects on which the current one depends.
Previously we were doing this computation in computeIfAbsent, which
meant we could end up with recursive calls to computeIfAbsent. This
"worked" under prior versions of the JDK, but is detected and reported
as a ConcurrentModificationException on JDK 9.