Allow binding an ExceptionMapper by an Exception type
This adds deterministic ExceptionMapper resolution, instead of relying
on the binding order of ExceptionMappers.

If an Exception is thrown without an explicit bound ExceptionMapper, the
superclasses of the Exception are explored for possible bound
ExceptionMappers. The closest bound superclass wins.

e.g If an application adds an explicit binding for NotFoundException,
that explicit Mapper is preferred over the base ActionExceptionMapper.
However, if no binding for BadRequestException exists, the base
ActionExceptionMapper is used instead.

Since an Exception always resolves to the same ExceptionMapper for the
life of a VM, it is safe to cache the resolution result.