Fix NotFoundAction handling
Previously POSTing to an unknown path would result in a 200 with an
empty response body, instead of the expected 404. Multiple problems:

1. The NotFoundAction was only bound for GET requests
2. The NotFoundAction only supported requests with Accept: text/plain

The NotFoundAction is now bound for both GET and POSTs, and supports
requests for any type of content (but always returns text/plain). Also
added a backstop in the WebActionsServlet to return a 404 if no suitable
action could be found; this handles cases where applications do not
install the NotFoundAction, or when a request arrives with an
unsupported HTTP method.