moshi handling for lists and maps of wire messages
Previously we were relying on moshi's default handling for repeated and
map fields on wire messages; this is fine for primitives, but populates
the collection with the wrong types when dealing with repeated or maps
of other wire messages. This adds proper support for dealing with those
types of collections.