Change DataSourcesConfig to inherit from LinkedHashMap
As a data class with a Map property, the DataSourcesConfig class
adds an additional level of nesting to configuration.

Eg:
```
data_sources:
  databases:
    exemplar:
      ...
    exemplar-readonly:
      ...
```

Rather than:
```
data_sources:
  exemplar:
    ...
  exemplar-readonly:
    ...
```

To achieve this DataSourcesConfig now implements the Map interface
such that when the Config object is being built from yaml the key-value
pairs within data_sources are interpreted as a Map on the root config.
(Jackson needs a concrete implementation of Map, which is why this is
done by inheriting from LinkedHashMap.)

The result is Config.data_sources is a Map<String, DataSourceConfig>,
but will still be bound by Guice as DataSourcesConfig.