SchemaMigrator for running and tracking schema migrations.
I intend for us to rely on external tools (Shift, gh-host) for production
schema changes. But for tests and development it's handy to have something
to apply schema changes.

This attempts a very small subset of what Flyway does. We don't need the
flexibility of Flyway and we benefit from additional control.