fail fast if any services are not Singleton
readiness checks will fail with a difficult to debug error if any
services are not Singletons (since the readiness check will get a new
instance, which will be in the NEW state). Check for this explicitly
and fail fast with an obvious error. Note this doesn't handle cases
where services are registered asEagerSingleton or bound toInstance.