Change tab domain from `/_admin/{tab}` to `/_tab/{tab}` (#384)
Why?: each tab used to compile to url domain `/_admin/{tab}`. Now that LoaderContainer is being used to render tabs, it should have control over routing their domain to the rendered tab, thus tab domains should be switched to `/_tab/{tab}` which is used for testing and for retrieving the tab compiled code.