Refactored clustering to DRY between k8s and fakes
The only difference between a Kubernetes cluster and a fake cluster is
the source for the events that change the cluster. Refactored clustering
into a DefaultCluster implementation that manages the ready member and
watch lists and computes changes to the cluster in response to incoming
events, and a KubernetesClusterWatcher which converts incoming k8s watch
events into the corresponding DefaultCluster changes.