Support for testing Services with external deps
Occasionally a Service requires access to an external resource during
startup. When testing those Services, we often use another Service to
startup that external resource within a test (cf
StartVitessService and StartEtcdService). For the application Service to
work properly, the test service that starts the external resource needs
to run first to make the external resource available. The best way to do
this is to have the application Service specify the test Service as a
dependency via DependentService.consumedKeys. ServiceTestingModule
provides a mechanism for doing this - you install a ServiceTestingModule
withExtraDependencies to specify your test specific dependencies, and
the module binds a Service that combines the application Service
dependencies with the extra dependencies to ensure that test specific
Services start before the application Service.