Push generic stuff to the JDBC level
This separates generic database interaction "stuff" like connection
pooling from Hibernate setup.

This accomplishes two things:
* Opens up for using other database interaction frameworks in the
  future.
* Forces us to build scalability sanity checks in such a way that they
  do not overuse Hibernate trickeries.

This is the first step towards extracting all this to a misk-jdbc module
which does not depend on misk-hibernate and should be able to stand on
its own.

Also includes a cross shard query detector as an example of what we can
build at this level.