Fix race condition in SqsJobQueueTest#stopsDeliveryAfterClose
When you close() a subscription, there may be one or more long-polls
outstanding. The AWS SDK does not give you a way to interrupt those
long-polls, so the threads remain listening for messages. For the test,
this means that it is still possible for the jobs enqueued after the
close to be delivered, since there might still be a thread in the
long-poll loop. To avoid this, we wait past the long poll timeout before
sending the post-close set of jobs