Lazy SessionFactory binding
Injecting a raw SessionFactory triggers an eager call to
SessionFactoryService.get() at injection time if you have a service that
(directly or indirectly) injects a Transacter. The SessionFactoryService
hasn't been started at that point which causes an IllegalStateException.
Using a Provider<SessionFactory> makes this call lazy which solves the
issue.