Add TestingResourceLoaderModule
Allows replacing ResourceLoaderModule in tests so filesystem: resources
are read from the classpath: instead. This allows tests to use a config
with a filesystem: resource and not rely on the underlying host
filesystem. A good example is an InjectorTest that verifies the guice
Injector can be created for the PRODUCTION enviornment.