Set useLegacyDatetimeCode=false jdbc property
This removes all the nonsense of attempting to convert between client
and server timezones. In prod this isn't really an issue since
everything run in UTC. However, for local development this can be a pain
since the local server is probably the developers timezone.

This isn't a problem for Vitess because it does not use j-connector.