Poll ZK for lease changes
With addition of listener support in #753, the ZK lease manager actually
now needs to poll for changes to leases in addition to receiving cluster
watch events. The node can receive cluster membership changes about
another member exiting the cluster, but before that member releases its
leases.

This change also changes the Lease check to be a bit nicer when there
are no members in the cluster. It will assume it doesn't own the lease
and log a warning, rather than logging an error and ignoring a potential
state transition. This should only have an effect on logs on service
startup/shutdown when there 0 services running.