Add support for providing trust_certificate_key_store_path and client_certificate_key_store_path
Some database drivers (e.g. Vitess) don't accept urls for certs in the
JDBC url they generate. So, for the sake of clarity, allow folks to
specify paths instead of urls. Fields ending in url imply that the string
begins with a scheme, so to ensure that clients do the right thing,
provide the path fields as an alternative.