Add module for binding AWS job queue handlers
Right now, to register job queue handlers, you need to define a separate
service that will in its `startUp()` function subscribe to queues, which
are actually consumed by the `SqsJobConsumer`. This is kind of a hairy
interface.

With `AwsSqsJobQueueModule`, you just have to install the module to bind
a handler to a queue.

I spent some time trying to write tests for it, but haven't found a good
way to do so because of how difficult it is to manipulate the service
dependency graph in tests. There isn't a good way for the service that
registers the subscriptions to depend on both `DockerSqs` startup and
queue creation. Since the module is pretty simple I'm just going to
leave it without tests.