Set key to file url instead of file name to allow config override files with the same names
For e.g., if we had the following override files when loading our test_app's misk config

/resources/db_config/test_app-config.yaml
/resources/cache_config/test_app-config.yaml

We would like to override our config with the content from both files. Currently, we only
use the filename as a key when creating a mapping of our config. This PR changes the key
to contain the entire URL.