Block consumers of SchemaMigratorService until vitess starts
StartVitessService takes care of running migrations in tests, however,
it does not produce the SchemaMigratorService key that other services
rely on, e.g SchemaValidatorService. This introduces a flake where the
validator might run before the migrator.

This required splitting the SchemaMigratorService into 2 services, one
that performs migrations in tests and one that validates migrations have
been run in prod envs.