Add ExternalDependency support to MiskTestExtension
We currently spin up external dependencies like Vitess and SQS in tests, by inserting additional services into the service dependency graph. It has forced us to use different workarounds, like ServiceTestingModule which modifies the service graph in tests, and StartVitessService which is a production dependency but really doesn't belong there.

@rhall had the idea that we could have the MiskTestExtension start up external dependencies instead, since they really are just test fixtures and are better handled by the testing framework.

It converts the DockerSqs service, which is simple but is a good enough demonstration.