Revert the previous attempt at fixing Vitess flaky tests
Now just always run StartVitessService, but only actually boot vitess if
it's needed (if vitess and in dev mode). This allows us to use
StartVitessService as a dependency to PingDatabaseService so we are sure
Vitess has been started before anything else. This starts the path of
moving Vitess to an ExternalDependency once that's ready.