Removed eager initalization
When running in STAGING or PRODUCTION environemnts, this object needs to
communicate with a live KMS, which makes it very difficult to test -
specifically the InjectorTest which loads each env's complete
configuration and makes sure Guice can initialize it.
Removing the `.asEagerSingleton()` part means that when the `Cipher`
object is constructed it won't need the `KmsClient` until first use.

There are other ways to exercise all the relevant code paths in `Cipher`
for testing without using a real KMS.