exposed TransacterService for db dependencies
Added an a public interface `TransactorService`, which service may
depend on if they require access to a database. Previously,
`SchemaMigratorService` was depended upon in order to ensure that all
services would start up in the right order.

`SessionFactoryService` remains a private class, which now implements
the `TransactorService` interface.

e.g.

If `MoviesService` previously depended on
`keyOf<SchemaMigratorService>(Movies::class)`, it may instead look
something like this:

```
install(ServiceModule<MoviesService>()
    .dependsOn<TransacterService>(Movies::class))
```

This hides the implementation detail of the presence of the schema
migrator, so that `MoviesService` (which only cares about a connection
to the database) may now only express a dependency on the database.