Change the way network interceptors work.
Today these are supposed to return a Response<ResponseBody> which is
written to the servlet's response stream at the end of the call. This
works for basic calls, but doesn't permit the action to access the
stream directly, such as via a @ResponseBody parameter.

This is step 2 or 3, the next step is a mechanical rename of the
following:
  - Request becomes HttpCall
  - FakeRequest becomes FakeHttpCall

Once these are done we can adopt the WebActionBinding API to properly
combine all binding features under a single system.