Release leases if the cluster weight is 0 (#1048)
This allows applications to run in an Active-Passive mode. If the
current weight for a cluster is 0, leases are not acquired or released
if already acquired. This allows an application to only run leased
background tasks in the Active region.

This iteration requires misk users to provide their own implementation
for ClusterWeightProvider. SQ services will use a global ZK impl, which
we could consider contributing as well in the future.