Use a RepeatedTaskQueue in SqsJobConsumer (#1071)
* Use a RepeatedTaskQueue in SqsJobConsumer

This fixes #1069, which was originally a bug in RepeatedTaskQueue back
in the day. We should prefer using RepeatedTaskQueue instead of hand
rolling threads to prevent repeating the same mistakes. In addition, job
consumers will backoff if there is no work or run into errors.

Now every parallel job consumer runs as a RepeatedTask and each Task
consumes N messages and dispatches them to separate executor service for
running.