Add capabilities to @Authenticated
This is the first part of https://github.com/cashapp/misk/issues/1078 to
add a new capabilities field to @Authenticated that will eventually
replace the roles field.

I did not Deprecate the roles field, because many repos have -WError set
which will fail the build if there is a Deprecated Warning.