Add QuerySniper and TransactionRunner
This change addresses issue (#230) and modifies the `Transacter` to:
 1. Run transactions with configurable time limits (entire transaction)
 2. Run individual queries with configurable time limits

Some new internal classes have been introduced to accomplish this:

 - `TransactionRunner`
 - `QuerySniper`

TransactionRunner
-----------------

Every transaction gets one. As its name suggests, it runs the
transaction block. It also schedules events to warn about or
cancel its transaction if the Timeouts are exceeded.

QuerySniper
-----------

Every session gets one. This is like `TransactionRunner`, but it listens
on individual hibernate queries.

How does this work?
-------------------

`Timeouts` are specified in the `DataSourceConfig`. The `HibernateModule`
will bind a `TimeoutsConfig` object based on the values configured
there.

There are three kinds of timeouts:
 - transaction timeouts: the time allotted for a transaction as a whole
 - query timeouts: the time allotted for a regular query
 - slow query timeouts: the time allotted for markedly slow queries

By default, these timeouts are all disabled.

When the timeouts are active, the runner attached to a transacter will
automatically kill long running transactions, and each session's sniper
will automatically kill long running queries.

You can opt out of a transaction's timeouts with

    transacter.noTimeouts().transaction { ... }

or specify custom transaction timeouts with

    transacter.withTimeouts(Timeouts.create(...)).transaction { ... }

Note that these timeouts do _NOT_ affect query timeouts, which must be
configured separately. It is possible to never have a transaction
timeout, but still have query timeouts, or vice versa.

If you know that a particular query is slow running, you can tell it
to use the slow query timeout configuration by

    transacter.transaction { session ->
      session.runSlowQuery { ... }
    }

The `runSlowQuery` lambda uses the session as a reciever.

If you want to change the query or slow query timeouts at runtime, you
can do so on the transacter:

    transacter.withQueryTimeouts(Timeouts.create(...))
         .withSlowQueryTimeouts(Timeouts.create(...))
         .transaction { ... }

That's it!