Fix flaky SqsJobQueueServiceTest
The test was flaky because it wasn't guaranteed that the JobConsumer had had a chance to run the job
in between the polls. The fix here is to add a CountDownLatch that'll wait until the job has had a
chance to run.

To test, I ran this test 300 times without failure.