Add an idempotenceKey to the JobQueue interface (#1252)
Add an idempotenceKey to the JobQueue interface

Most likely the underlying queueing system (e.g SQS) supports at least
once delivery of a message. Consumers are going to need an idempotence
key in order to filter duplicate messages.

The eventual goal is to require the idempotence key, but for backwards
compatibility a random token is generated for messages without one. Once
we upgrade all clients to the new methods (and deploy the changes) we
can remove the old methods.

For the SQS impl, the idempotence key is added as a message attribute.