Query timeout
We need a client side query timeout to handle Aurora instances or
vttablets that are hard killed.

For Aurora if an instance is hard killed it won't ever respond to the
query which will hold up threads in the http thread pool causing an
extended outage.

For Vitess the query sniper runs in the vttablet so it will handle an
Aurora instance being hard killed but if the vttablet itself is hard
killed the query sniper won't run again holding up threads and causing
extended outages.

1 minutes is quite a long timeout. If you want shorter timeouts they
can be overrided at the query level using the JPA query hint
javax.persistence.query.timeout.

This timeout is only applied to Hibernate queries. For raw JDBC queries
we will need something else.