Add idempotence key and jobqueue internal metadata attributes
This changes brings internal and misk jobqueue libraries closer in
feature-parity by offering the option to specify an application assigned
`idempotence_key` (as opposed to the system assigned `id`, only available after
a job is enqueued), which can be used by `JobHandler` implementations to detect
duplicate jobs.

Internal attributes required for jobqueue-related functionality are stored in a
single message attribute (JSON-encoded `Map<String, String>`), leaving a
maximum of 8 arbitrary attributes for applications (the other reserved
attribute is for trace id; while not always present, it makes sense to assume
it'll exist and just consistently limit app keys to 8).

This change does not introduce dependency on metadata or any of its attributes,
meaning new jobs will begin having these properties as services are deployed.

The two metadata attributes added with this change are:

* Idempotence key: our internal jobqueue libs have always provided idempotence
key support, which some services heavily rely on to avoid duplicate work.

* Origin queue: when a job is transferred from main or retry queue to
dead-letter queue, it does not contain any metadata about its origins.
As we prepare to roll out a(n optional) more functional solution for
dead-lettered job management, which relies on a single, global dead-letter
queue, it's necessary to record the main queue where a job came from (so that
it's known where it should be sent to, in case of reprocessing).

Support for global DLQ (vs per-queue DLQ) will be introduced in a follow-up PR.