Allow no queueing in Jetty
The QueuedThreadPool first submits all requests to the queue and the
thread pool pulls from the queue. With this impl, there is no way to
disable queueing on the server.

If config sets queue size to 0, a different impl is used which direclty
hands the reques to the ThreadPool and rejects the request if max
threads are in use.