Add a dedicated Connector for health checks
This dedicated Connector uses a dedicated thread pool for health
checking. If the Server is currently queueing/rejecting requests, it can
still process liveness/readiness checks from k8s.

Previously a queued health check could result in k8s killing the
server, even if it was perfectly healthy and serving traffic. This could
result in cascading failures since the traffic would shift to another
node, creating a queue and killing that container.

To use the experimental feature, apps set the health_port to 8080 and
the port to something else, like 8081 for local development.