Do not update all connection metrics when a connection is removed
If there are a lot of active connections this can cause a lot of
contention on the metrics, since you might have many closing connections
all trying to update the same metrics.

There is no need to refresh the metrics when closing a connection. We
can just save the stats of the closed connections so they are included
in the next call to refereshMetrics.