Do not update all connection metrics when a connection is removed (#1542)
* Revert "Only update connection metrics in a background thread (#1541)"

This reverts commit 899d76e4712401d7fdb923ccdec01a324bb0ec9e.

* Do not update all connection metrics when a connection is removed

If there are a lot of active connections this can cause a lot of
contention on the metrics, since you might have many closing connections
all trying to update the same metrics.

There is no need to refresh the metrics when closing a connection. We
can just save the stats of the closed connections so they are included
in the next call to refereshMetrics.