Do not update all connection metrics when a connection is removed (#1542)
* Revert "Only update connection metrics in a background thread (#1541)"

This reverts commit cfdb9d5b5671231229db9eb4aef9f9e565203ab2.

* Do not update all connection metrics when a connection is removed

If there are a lot of active connections this can cause a lot of
contention on the metrics, since you might have many closing connections
all trying to update the same metrics.

There is no need to refresh the metrics when closing a connection. We
can just save the stats of the closed connections so they are included
in the next call to refereshMetrics.