Remove QueryTracingListener (#1546)
This class is buggy and we often see log lines like
`lastScope db-insert wasn't closed` emitted by this class. That
indicates it didn't close spans, presumably due to a hook not running
when it should have.

We use Datadog now which has automatic instrumentation for both
Hibernate and JDBC. So rather than trying to fix, let's just get rid of
this class.