Extend mockito extension with more helper functions
It appears there are some rough edges with using Mockito with Kotlin. For example, if you use Mockito matchers directly in Kotlin, they typically return null which will trigger an NPE. The work around is to use helper functions. The real answer seems deeper philosophically, like Kotlin having its own test libraries.

Also removing the extension of Any for the whenever helper function because otherwise you can't stub out methods that return a nullable.

See https://stackoverflow.com/questions/30305217/is-it-possible-to-use-mockito-in-kotlin