Load classpath resources lazily
Previously we used Guava's classpath scanner to load all resources
eagerly. This was easy but it also required us to load a lot of
resources that we might not need, potentially slowing app launch.

Also this should make it possible to make further improvements to
load resources in more complex classloader situations.