Add mechanism to add additional MDC context to web requests
Description
-----------
Applications owners want to be able to add additional contextual
information in the MDC in such a way that gets included into the
`ExceptionHandlingInterceptor` and `RequestLoggingInterceptor`.

Add a way to enrich the `RequestLogContextInterceptor` by having a
MapBinder where the key is the MDC key and the value will be returned by
a LogContextProvider.