Revreted to using TypeLIteral to bind a generic type
Apparently, some incopatibility between Kotlin and Java doesn't like the
"kotlinized" version of the bind statement when it used with a generic type.

So Guice sees `bind<Map<K,V>().annotatedWith<SomeAnnotation>()` as `@SomeAnnotation Map` instead of `@SoneAnnotation Map<K,V>`
I decided to resolve this issue by using a more Java-esque syntax to explicitly tell the bind statement that we're trying to bind a Map with specific generic types.