Allow better configuration over the set of SSL protocols used. (#1792)
The default codepath allows some weak ciphers, and for certain applications
(in particular ones that don't have to deal with older client software) it's
nice to have an option to just use strong modern protocols. This commit
adopts the compatible/modern/restricted terminology from Google's cloud
platform docs and allows users to configure which set they want to use.
The existing behaviour (which is now the "compatible" policy) is unchanged
with the exception of also allowing TLS v1.3 and remains the default
option.