Introduce the Quota-Path header for better concurrency limiting
If a single action is used for different types of work this header can
be provided by callers to bucket work by kind. This allows the concurrency
limiter to make better predictions of timeouts and should increase overall
availability for endpoints that handle heterogeneous traffic.

The initial use case for this is internal endpoints that handle different
event topics.