Revert part of #1958 to try and fix shutdown deadlock (#1976)
The change to the synchronization lock object in #1958 appears
to have resulted in an intermittent deadlock. Tests that exercise
this code heavily frequently hit the deadlock and time out.
Details of the thread stacks during the deadlock can be found
in issue #1975.

This patch reverts the change to the synchronization lock object,
which appears incidental to the main change that was being made
in #1958. Hopefully this will fix the deadlock while retaining
the intended functional change. Follow-up PRs can be used to
implement a better solution.

Fixes #1975.