Revisit some api/implementation dependencies (#1973)
The starting point for this was looking at a sensible set of dependencies for the sample project. This led to two changes:

- misk should have an api dep on misk-inject because MiskApplication's API uses classes from it (and misk is heavily tied to misk-inject and Guice)
- misk-actions should have an api dep on okhttp because the Actions API uses classes from it, like Headers

The other dependencies in the exemplar projects were redundant; Misk already included them.

Need to cut a release before changing the references in the exemplar from projects to external references. Having external references makes the samples copy-pasteable.

Co-authored-by: Chris Ryan <74329397+chris-ryan-square@users.noreply.github.com>