Support Customization of gRPC status by mappers
Instead of always mapping an HTTP status code to a gRPC status, give ExceptionMappers a chance to
explicitly produce a gRPC status and message. This gives us more fine-grained control over exactly
what gets returned in :grpc-message headers.

In particular, we want to return the raw exception message in gRPC responses while still masking it
for HTTP responses. This is because gRPC clients are generally internal, while HTTP clients may be
external organizations that we don't want to leak internal information to.