Log on properties present in config yaml but not in object
Why: services may introduce new config before a new version of the service is deployed.
This new config should be harmless to the service until it needs the config.

This is a change in behaviour - previously new properties present in the config yaml
but not in the object will throw an error.

This also adds additional safeguards for when properties are misspelled: they would
fail because the property is required but missing, and checks for similar properties
to help identify a spelling mistake.