Add google.rpc.Status-compatible gRPC errors (#2126)
Including a full un-padded-base64-proto-encoded copy of Status as a grpc-status-details-bin trailer is the standard way to communicate rich errors to clients in gRPC. Unfortunately, this detail didn't make it in the spec (yet).

It's widely supported in practice. grpc-go has it, grpc-java, as well as Python, Ruby and C++ and virtually every compatible implementation.

It lets servers return `google.rpc.Status`-encoded errors with embedded custom details (for example `error_details.proto`). In Misk we do so by throwing a `WebApiException`. To keep the API compatible with Wire-generated messages, `WebApiException` accepts a list of `AnyMessage`—a Wire runtime. This makes it easy to use but it means we need to compile our own copy of `status.proto`.

Reference:
https://github.com/googleapis/googleapis/blob/master/google/rpc/status.proto