Don't publish enforcedPlatform dependencies to downstream consumers
In general this seems to be bad practice, to the point where gradle 7.2
disallows it by default. Misk currently requires all downstream consumers
to use the exact version of platform dependencies specified. Instead,, misk
should just indicate the version it wants and allow downstream consumers
to resolve the dependency to that (or possibly a newer) version.