feat: AccessInterceptor: allow >1 access annotations on actions.
Consider an action that should be accessible by multiple groups.
A natural way to express this is:

```kt
@Group1Access
@Group2Access
@Get("/some-action")
fun someAction() {
  // ...
}
```

Currently, Misk does not support this. It is hard-coded to throw an
IllegalStateException whenever an action is annotated with more than
one access annotation. As far as I can tell, this has been the case
forever. This limitation forces Misk users to resort to unfortunate
workarounds, such as defining a @Group1OrGroup2Access annotation that
manually aggregates the underlying services/capabilities.

This PR changes AccessInterceptor to allow multiple access
annotations. This is fairly easy because the class already represents
the allowed services/capabilities as sets. All that is needed is to
aggregate the services/capabilities from all the access annotations
instead of just one.

(As a special case, this PR makes it an error to specify
@Unauthenticated alongside other access annotations, since this is a
logical contradiction.)