Add GoogleSpannerModule to misk-gcp
This commit adds support for loading, configuring, using, and testing a Google Spanner client in Misk services. It makes Spanner feel more like a standard datastore in Misk, like DynamoDB.

It contains code to download an emulator from Docker and run it for tests in both local development and CI for convenience, as well as utilities for clearing the DB in between tests.

Lastly, there is now documentation written on how to use the GoogleSpannerModule. It's not perfect, but it's a start and should be useful to engineers on my team that are using it.