Don't publish enforcedPlatforms to downstream consumers
When an enforcedPlatform is used in the dependency list,
it forces all downstream consumers of that artifact to use the
exact same version that is specified. This is generally not
desirable, and in fact gradle 7.2 produces the following error:

```
> Invalid publication 'maven':
    - Variant 'apiElements' contains a dependency on enforced platform 'io.grpc:grpc-bom'
    - Variant 'runtimeElements' contains a dependency on enforced platform 'io.grpc:grpc-bom'
  In general publishing dependencies to enforced platforms is a mistake: enforced platforms shouldn't be used for published components because they behave like forced dependencies and leak to consumers. This can result in hard to diagnose dependency resolution errors.
```

Although the error can be suppressed if we are doing this
intentionally, I believe we are NOT doing this intentionally,
and so instead we should avoid using enforcedPlatform, and use
platform instead.