Give Limiter a nanosecond clock like it expects
The `Supplier<Long> clock` that AbstractLimiter.Builder.clock() expects is supposed to return a
_nanosecond_ value, not a millisecond value. All of the internal machinery of concurrency-limits
assumes the clock values are in nanoseconds, so we should give it what it expects.

I'm not sure this actually changes any behavior, because VegasLimiter is mostly just dividing rtt
against min_rtt ("rtt_noload"), so changing the scale of both values shouldn't affect anything.

There doesn't seem to be a good way to test this, as I'm not even sure it affects observable
behavior. If it does have an effect, it would be in the rate at which VegasLimiter reacts, which
would be difficult to test reliably.