Make counter help string not nullable (#2272)
I was in this code for a different change and I noticed that the help string for counter was nullable, but the help string for all other metric types was not.

I was curious as to whether that was intensional, so I tracked the code and discovered that even though it's marked as nullable, the code actually fails if the value is null or empty.

https://github.com/prometheus/client_java/blob/master/simpleclient/src/main/java/io/prometheus/client/SimpleCollector.java#L170

This change elevates the failure from a runtime failure to a compile-time failure which is preferable. Strictly speaking, this is not a backward compatible change (because an optional field is now no longer optional) but reading the code, all of those usages that didn't set this parameter should have thrown an IllegalArguement exception so I don't think there will be any valid cases that won't work.