misk-actions depends on newly created misk-action-scopes
A recent PR, pulled the ActionScoped class out of the misk-actions subproject into a new misk-action-scopes subproject. This was a breaking change for any repositories having a dependency on misk-actions and using ActionScoped.
https://github.com/cashapp/misk/pull/2301

This commit adds an api dependency in misk-actions to misk-action-scopes, so that the misk-actions package included the classes from that dependency.